+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Version 2.0.4 (2022-10-25)
 #. ``sc.stackedbar()`` will automatically plot a 2D array as a stacked bar chart.
 #. ``sc.parallelize()`` now always tries ``multiprocess`` if an exception is encountered and ``die=False`` (unless ``parallelizer`` already was ``'multiprocess'``).
 #. Added a ``die`` argument to ``sc.save()``.
+#. Added a ``prefix`` argument to ``sc.urlopen()``, allowing e.g. ``http://`` to be omitted from the URL.
 
 
 Version 2.0.3 (2022-10-24)